requirements.txt
Python==3.12.7
Django==5.1.2
django-environ==0.11.2
djangorestframework==3.15.2
mysqlclient==2.2.5
pillow==11.0.0
django-filter==24.3
drf-yasg==1.21.8
django-silk==5.3.1
djangorestframework-simplejwt==5.3.1